requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.8
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
beartype==0.11.0
bitrix24-rest==1.1.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
fast-bitrix24==1.5.11
Flask==2.2.2
frozenlist==1.3.1
icontract==2.6.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
more-itertools==8.14.0
multidict==6.0.2
requests==2.28.1
six==1.16.0
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
